requirements.txt
discord.py 
yt-dlp 
PyNaCl
ffmpeg